Migrate existing data with US as a default country
As a side-effect this will add the US as a country in any instance's
database, even if there *aren't* existing campaigns, but for the time
being that seems to be okay (and much easier to implement). We can
assume that the list of countries in the database maps 1-to-1 with the
countries implemented in call_server/political_data.