Update campaign forms to fill contents based on type
This change moves the selection of campaign type back a level, to the
new country form. This way we can populate the region and subtype
options based on the campaign type. This isn't ideal, but it's simple :)

Because of this change, the campaign form no longer asks the user to
pick a region unless one is actually needed for the campaign type. In
addition, only subtypes for the selected type are shown (if any).

In the future we may want to bring these forms back together with a bit
of JavaScript.